Crate mbpr

Nom based decoder for Memcached Binary Packet Protol

Supports all current opcodes. The goal of this interface is simplicity and speed. There is very little validation done on packet construction the expectation of that is placed on higher libraries.

The parsing is fairly strict.

For examples please see tests directory or keep reading.

This was created based on Memcached Wiki

Structs

Encoder
Data type used to encode data efficient

OwnedRequest
Clones Buffers

OwnedResponse
Clones Buffers

ReqHeader
Memcached Request Packet Header

Request
Memcached Request Packet

ResHeader
Memcached Response Packet Header

Response
Memcached Response Packet

Enums

Fault
Parsing Error structure

OpCode
Memcache Opcodes

ParseResult
Internal Result type

StatusField
Status Field

Traits

Encoding
Trait for encoding the value into a packet

PacketVal
Values encoded within Memcached packets